3.4.2.0

+++ Geometry/GeometryTest/TestFasterPolyShape.cpp
@@ -481,7 +481,7 @@ void CTestFasterPolyShape::TestIShapeOpen()
    pnts[1].Release();
    pnts[2].Release();
    pnts[3].Release();
-   Enum->Next(nPoints,&pnts[0],&fetched);
+   Enum->Next((ULONG)nPoints,&pnts[0],&fetched);
    TRY_TEST(fetched,nPoints);
 
    pnts[0]->get_X(&x);
@@ -578,7 +578,7 @@ void CTestFasterPolyShape::TestIShapeOpen()
    pnts[2].Release();
    pnts[3].Release();
    pnts[4].Release();
-   Enum->Next(nPoints,&pnts[0],&fetched);
+   Enum->Next((ULONG)nPoints,&pnts[0],&fetched);
    TRY_TEST(fetched,nPoints);
 
    pnts[0]->get_X(&x);
