+++ Geometry/GeometryTest/TestUBeam2.cpp
@@ -420,7 +420,7 @@ void CTestUBeam2::TestIShape()
    coll->get__Enum(&Enum);
    CComPtr<IPoint2d> points[17];
    ULONG fetched;
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    Float64 x,y;
@@ -615,7 +615,7 @@ void CTestUBeam2::TestIShape()
    TRY_TEST( cPoints, 6 );
 
    coll->get__Enum(&Enum);
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    points[0]->get_X(&x);
@@ -670,7 +670,7 @@ void CTestUBeam2::TestIShape()
    TRY_TEST( cPoints, 12);
 
    coll->get__Enum(&Enum);
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    points[0]->get_X(&x);
@@ -775,7 +775,7 @@ void CTestUBeam2::TestIShape()
 
 
    coll->get__Enum(&Enum);
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    points[0]->get_X(&x);
@@ -852,7 +852,7 @@ void CTestUBeam2::TestIXYPosition()
    coll->get__Enum(&Enum);
    CComPtr<IPoint2d> points[17];
    ULONG fetched;
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    Float64 x,y;
@@ -965,7 +965,7 @@ void CTestUBeam2::TestIXYPosition()
    TRY_TEST( cPoints,17);
 
    coll->get__Enum(&Enum);
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    points[0]->get_X(&x);
@@ -1232,7 +1232,7 @@ void CTestUBeam2::TestIXYPosition()
    TRY_TEST( cPoints,17);
 
    coll->get__Enum(&Enum);
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    points[0]->get_X(&x);
